Move all logic out of init
