+++ service_standard_python/middlewares.py
@@ -1,0 +1,12 @@
+from prometheus_client import make_wsgi_app
+from werkzeug.middleware.dispatcher import DispatcherMiddleware
+
+
+def add_prometheus_middleware(app):
+    app.wsgi_app = DispatcherMiddleware(app.wsgi_app, {
+        '/metrics': make_wsgi_app()
+    })
+
+
+def make_prometheus_wsgi_app():
+    return make_wsgi_app()
